requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
firebase-admin==4.5.3
google-api-core==1.31.3
google-api-python-client==2.24.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
httplib2==0.20.1
idna==3.2
lxml==4.6.3
msgpack==1.0.2
packaging==21.0
ply==3.11
proto-plus==1.19.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyshorteners==1.0.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
stone==3.2.1
tornado==6.1
tzlocal==3.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
